Privacy Policy.md
Privacy Policy

This privacy policy sets out how Bubble Block Puzzle uses and protects any information that you give Bubble Block Puzzle when you use this mobile games. Bubble Block Puzzle is comitted ensuring that your privacy is protected.Should we ask you to provide certain information by which you can be identiified when using this mobile game developed by us,then you can be assured that it will only be used accordance with this privacy statement. Bubble Block Puzzle may change this policy from time to time by updating this page.You should check this page from time to time to ensure that you are happy with any changes.This policy is effective from 07/02/2017.

What we collect

We do not collect any information if you download and play our games.If you write feedback to us , we will collect. -You email address

What we do with the information we gather

We require this information to communicate with your request/feedback and provide you with a better service,and in particular for the following reasons: -Internal record keeping. -We may use the information to improve our products and services. -We may periodically send promotional emails about new products,special offers or other information which we think you may find interesting using the email address which you have provided. -From time to time ,we may also use your information to contact you for market research purposes.We may contact you by email.

Security

We are committed to ensuring that your information in secure.In order to prevent unauthorised access or disclosure ,we have put in place suitable physical,electronic and managerial procedures to safeguard and secure the information we collect online.